@@ -5,7 +5,7 @@
   "main": "dist/draw2d.js",
   "scripts": {
     "build": "webpack --env dev && webpack --env build",
-    "dev": "webpack --progress --colors --watch --env dev"
+    "watch": "webpack --progress --watch --env dev"
   },
   "repository": {
     "type": "git",
@@ -24,24 +24,26 @@
   },
   "homepage": "https://github.com/freegroup/draw2d",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.0.3",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-env": "^1.6.1",
-    "css-loader": "^0.28.11",
-    "eslint": "^4.13.1",
-    "eslint-loader": "^2.0.0",
-    "glob": "^7.1.3",
+    "@babel/cli": "^7.20.7",
+    "@babel/core": "^7.20.12",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-syntax-logical-assignment-operators": "^7.10.4",
+    "@babel/preset-env": "^7.20.2",
+    "babel-loader": "^9.1.2",
+    "babel-plugin-add-module-exports": "^1.0.4",
+    "babel-preset-modern-browsers": "^15.0.2",
+    "css-loader": "^6.7.3",
+    "eslint-webpack-plugin": "^3.2.0",
+    "glob": "^8.0.3",
     "script-loader": "^0.7.2",
-    "style-loader": "^0.21.0",
-    "webpack": "^4.0.0",
-    "webpack-cli": "^3.0.8",
-    "yargs": "^10.0.3"
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "yargs": "^17.6.2"
   },
   "dependencies": {
     "canvg-browser": "1.0.0",
-    "shifty": "2.3.1"
+    "shifty": "2.20.0"
   }
 }
