@@ -151,7 +151,7 @@ draw2d.Canvas = Class.extend(
       this.commandStack.addEventListener(function (event) {
         if (event.isPostChangeEvent() === true) {
           _this.calculateConnectionIntersection()
-          _this.linesToRepaintAfterDragDrop.each(function (i, line) {
+          _this.linesToRepaintAfterDragDrop.each((i, line) => {
             line.svgPathString = null
             line.repaint()
           })
@@ -885,16 +885,16 @@ draw2d.Canvas = Class.extend(
       }
 
       if (figure instanceof draw2d.shape.basic.Line) {
-        this.lines.add(figure)
         this.linesToRepaintAfterDragDrop = this.lines
+        this.lines.add(figure)
       } else {
         this.figures.add(figure)
-        if (typeof y !== "undefined") {
-          figure.setPosition(x, y)
-        } else if (typeof x !== "undefined") {
-          figure.setPosition(x)
-        }
       }
+      
+      if (typeof x !== "undefined") {
+        figure.setPosition(x,y)
+      }
+
       figure.setCanvas(this)
 
       // to avoid drag&drop outside of this canvas
@@ -922,7 +922,7 @@ draw2d.Canvas = Class.extend(
       // of the connections
       if (figure instanceof draw2d.shape.basic.PolyLine) {
         this.calculateConnectionIntersection()
-        this.linesToRepaintAfterDragDrop.each(function (i, line) {
+        this.linesToRepaintAfterDragDrop.each((i, line) => {
           line.svgPathString = null
           line.repaint()
         })
