@@ -1,6 +1,4 @@
 import draw2d from 'packages'
-import extend from "./util/extend";
-
 
 /**
  * @class
@@ -27,21 +25,21 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
       this.name = null
       this.coronaWidth = 5 // the corona width for the hitTest method. Useful during drag&drop of ports. Better SnapTo behavior.
 
-      this._super(extend({
+      this._super({
           bgColor: "#4f6870",
           stroke: 1,
           diameter: 10,
           color: "#1B1B1B",
           selectable: false
-        }, attr),
-        extend({
+          , ...attr},
+        {
           coronaWidth: this.setCoronaWidth,
-          semanticGroup: this.setSemanticGroup
-        }, setter),
-        extend({
+          semanticGroup: this.setSemanticGroup,
+          ...setter},
+        {
           coronaWidth: this.getCoronaWidth,
-          semanticGroup: this.getSemanticGroup
-        }, getter))
+          semanticGroup: this.getSemanticGroup,
+          ...getter})
 
 
       // status vars for user interaction
@@ -83,7 +81,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
       this.installEditPolicy(new draw2d.policy.port.IntrusivePortsFeedbackPolicy())
       // this.installEditPolicy(new draw2d.policy.port.ElasticStrapFeedbackPolicy());
 
-      // a port handles the selection handling always by its own regardless if
+      // a port handles the selection handling always by its own, regardless if
       // the port is part of an composite, node, group or whatever.
       this.portSelectionAdapter = () => this
     },
@@ -95,7 +93,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
     /**
      *
      * set the maximal possible count of connections for this port.<br>
-     * This method din't delete any connection if you reduce the number and a bunch of
+     * This method didn't delete any connection if you reduce the number and a bunch of
      * connection are bounded already.
      *
      * @param {Number} count the maximal number of connection related to this port
@@ -159,11 +157,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
      **/
     setConnectionAnchor: function (anchor) {
       // set some good defaults.
-      if (typeof anchor === "undefined" || anchor === null) {
-        anchor = new draw2d.layout.anchor.ConnectionAnchor()
-      }
-
-      this.connectionAnchor = anchor
+      this.connectionAnchor = anchor ?? new draw2d.layout.anchor.ConnectionAnchor()
       this.connectionAnchor.setOwner(this)
 
       // the anchor has changed. In this case all connections needs an change event to recalculate
@@ -274,7 +268,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
      * @returns {this}
      **/
     setBackgroundColor: function (color) {
-      this.lighterBgColor = new draw2d.util.Color(color).lighter(0.3).rgba()
+      this.lighterBgColor = new draw2d.util.Color(color).lighter(0.3).hash()
       this._super(color)
 
       return this
@@ -284,7 +278,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
      *
      * Set a value for the port. This is useful for interactive/dynamic diagrams like circuits, simulator,...
      *
-     * @param {Object} value the new value for the port
+     * @param {Object} value the new value for the port. Can be "undefined" or "null"
      * @returns {this}
      */
     setValue: function (value) {
@@ -294,9 +288,9 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
       }
       let old = this.value
       this.value = value
-      if (this.getParent() !== null) {
-        this.getParent().onPortValueChanged(this)
-      }
+
+      this.getParent()?.onPortValueChanged(this)
+      
       this.fireEvent("change:value", {value: this.value, old: old})
 
       return this
@@ -320,8 +314,7 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
         return
       }
 
-      attributes = attributes || {}
-
+      attributes ??= {}
 
       // a port did have the 0/0 coordinate in the center and not in the top/left corner
       //
@@ -375,15 +368,11 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
      * @private
      */
     setParent: function (parent) {
-      if (this.parent !== null) {
-        this.parent.off(this.moveListener)
-      }
+      this.parent?.off(this.moveListener)
 
       this._super(parent)
 
-      if (this.parent !== null) {
-        this.parent.on("move", this.moveListener)
-      }
+      this.parent?.on("move", this.moveListener)
     },
 
 
@@ -432,8 +421,6 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
         return false
       }
 
-      let _this = this
-
 //        this.getShapeElement().insertAfter(this.parent.getShapeElement());
       // don't call the super method. This creates a command and this is not necessary for a port
       this.ox = this.x
@@ -443,11 +430,11 @@ draw2d.Port = draw2d.shape.basic.Circle.extend(
 
       // notify all installed policies
       //
-      this.editPolicy.each(function (i, e) {
+      this.editPolicy.each( (i, e)=>  {
         if (e instanceof draw2d.policy.figure.DragDropEditPolicy) {
           // DragStart operation can send a veto for the dragStart
           // @since 6.1.0
-          canStartDrag = canStartDrag && e.onDragStart(_this.canvas, _this, x, y, shiftKey, ctrlKey)
+          canStartDrag = canStartDrag && e.onDragStart(this.canvas, this, x, y, shiftKey, ctrlKey)
         }
       })
 
