@@ -1,6 +1,4 @@
 import draw2d from 'packages'
-import extend from 'util/extend'
-
 
 /**
  * @class
@@ -33,7 +31,7 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
 
     this.strokeScale = true // scale the stroke width of the children nodes if the parent resize
 
-    this._super(extend({stroke: 0, bgColor: null}, attr), setter, getter)
+    this._super({stroke: 0, bgColor: null, ...attr}, setter, getter)
   },
 
   /**
@@ -69,12 +67,12 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
     }
 
     if (this.cssClass === null) {
-      this.svgNodes.forEach(function (e) {
+      this.svgNodes.forEach( (e) => {
         e.node.removeAttribute("class")
       })
     }
     else {
-      this.svgNodes.forEach(function (e) {
+      this.svgNodes.forEach( (e) => {
         e.node.setAttribute("class", cssClass)
       })
     }
@@ -103,7 +101,7 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
       this.scaleY = this.height / this.originalHeight
     }
 
-    attributes = attributes || {}
+    attributes ??= {}
 
     this.applyAlpha()
 
@@ -204,9 +202,7 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
 
       // and all inner children
       //
-      if (this.svgNodes !== null) {
-        this.svgNodes.toFront()
-      }
+      this.svgNodes?.toFront()
 
       if (this.canvas !== null) {
         let figures = this.canvas.getFigures()
@@ -300,7 +296,7 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
 
     // bring all children figures in front of the parent
     // run reverse to the collection to care about the z-order of the children)
-    this.children.each(function (i, child) {
+    this.children.each( (i, child) => {
       child.figure.toBack(figure)
     }, true)
 
@@ -325,15 +321,14 @@ draw2d.SetFigure = draw2d.shape.basic.Rectangle.extend(
 
     // and last but not least - the ports are always on top
     //
-    let _this = this
-    this.getPorts().each(function (i, port) {
-      port.getConnections().each(function (i, connection) {
-        connection.toFront(_this)
+    this.getPorts().each((i, port) => {
+      port.getConnections().each((i, connection) =>  {
+        connection.toFront(this)
       })
       // a port should always be in front of the shape doesn't matter what the
       // "figure" parameter says.
       //
-      port.toFront(_this)
+      port.toFront(this)
     })
 
     return this
