@@ -1,6 +1,4 @@
 import draw2d from '../../packages'
-import extend from '../../util/extend'
-
 
 /**
  * @class
@@ -68,13 +66,7 @@ draw2d.layout.connection.InteractiveManhattanConnectionRouter = draw2d.layout.co
 
   onInstall: function (conn) {
     conn.installEditPolicy(new draw2d.policy.line.OrthogonalSelectionFeedbackPolicy())
-    if (!conn._routingMetaData) {
-      conn._routingMetaData = {
-        routedByUserInteraction: false,
-        fromDir: -1,
-        toDir: -1
-      }
-    }
+    conn._routingMetaData ??= { routedByUserInteraction: false,fromDir: -1,toDir: -1}
   },
 
   onUninstall: function (conn) {
@@ -117,7 +109,7 @@ draw2d.layout.connection.InteractiveManhattanConnectionRouter = draw2d.layout.co
     let max = Math.max
     let min = Math.min
 
-    routingHints = routingHints || {oldVertices: new draw2d.util.ArrayList()}
+    routingHints ??= {oldVertices: new draw2d.util.ArrayList()}
     let oldVertices = routingHints.oldVertices
     let vertexCount = oldVertices.getSize()
 
@@ -126,7 +118,6 @@ draw2d.layout.connection.InteractiveManhattanConnectionRouter = draw2d.layout.co
 
     let toPt = conn.getEndPosition()
     let toDir = conn.getTarget().getConnectionDirection(conn.getSource())
-
 
     // the port orientation has been changed. This can happen if the node rotates. In this case
     // we must recalculate the routing.
@@ -525,10 +516,10 @@ draw2d.layout.connection.InteractiveManhattanConnectionRouter = draw2d.layout.co
   getPersistentAttributes: function (line, memento) {
     memento.vertex = []
 
-    line.getVertices().each(function (i, e) {
+    line.getVertices().each( (i, e) =>{
       memento.vertex.push({x: e.x, y: e.y})
     })
-    memento.routingMetaData = extend({}, line._routingMetaData)
+    memento.routingMetaData = { ...line._routingMetaData}
 
     return memento
   },
@@ -549,17 +540,10 @@ draw2d.layout.connection.InteractiveManhattanConnectionRouter = draw2d.layout.co
       line.lineSegments = new draw2d.util.ArrayList()
 
       line.setVertices(memento.vertex)
-
-      /*
-       line.vertices     = new draw2d.util.ArrayList();
-       $.each(memento.vertex, function(i,e){
-       line.addPoint(e.x, e.y);
-       });
-       */
     }
 
     if (typeof memento.routingMetaData !== "undefined") {
-      line._routingMetaData = extend({}, memento.routingMetaData)
+      line._routingMetaData = {...memento.routingMetaData}
     }
   }
 
