@@ -1,5 +1,4 @@
 import draw2d from '../../packages'
-import extend from '../../util/extend'
 
 draw2d.SnapToHelper = {}
 
@@ -42,16 +41,17 @@ draw2d.policy.canvas.SnapToEditPolicy = draw2d.policy.canvas.CanvasPolicy.extend
     this.lineColor = null
 
     this._super(
-      extend({
-        lineColor: "#51C1FC"
-      }, attr),
-      extend({
+      {
+        lineColor: "#51C1FC",
+        ...attr
+      },
+      {
         // @attr {draw2d.util.Color} color the line color of the snapTo lines */
-        lineColor: this.setLineColor
-      }, setter),
-      extend({
-        lineColor: this.getLineColor
-      }, getter))
+        lineColor: this.setLineColor,
+        ...setter},
+      {
+        lineColor: this.getLineColor,
+        ...getter})
   },
 
   /**
