@@ -32,7 +32,7 @@ draw2d.shape.basic.Text = draw2d.shape.basic.Label.extend(
     init: function (attr, setter, getter) {
       this.cachedWrappedAttr = null
 
-      this._super(extend({width: 100, height: 50, resizeable: true}, attr), setter, getter)
+      this._super({ width: 100, height: 50, resizeable: true,...attr}, setter, getter)
 
       this.installEditPolicy(new draw2d.policy.figure.WidthSelectionFeedbackPolicy())
     },
@@ -47,7 +47,7 @@ draw2d.shape.basic.Text = draw2d.shape.basic.Label.extend(
       }
 
       // style the label
-      this.svgNodes.attr(extend({}, this.calculateTextAttr(), this.wrappedTextAttr(this.text, this.getWidth() - this.padding.left - this.padding.right)))
+      this.svgNodes.attr({...this.calculateTextAttr(), ...this.wrappedTextAttr(this.text, this.getWidth() - this.padding.left - this.padding.right)})
 
       // set of the x/y must be done AFTER the font-size and bold has been set.
       // Reason: the getHeight method needs the font-size for calculation because
@@ -101,12 +101,10 @@ draw2d.shape.basic.Text = draw2d.shape.basic.Label.extend(
       if (this.cachedMinWidth === null) {
         // get the longest word in the text
         //
-        let longestWord = this.text.split(" ").reduce(function (arg1, arg2) {
-          return arg1.length > arg2.length ? arg1 : arg2
-        })
+        let longestWord = this.text.split(" ").reduce( (arg1, arg2)=>arg1.length>arg2.length?arg1:arg2)
         let svgText = this.canvas.paper
           .text(0, 0, longestWord)
-          .attr(extend({}, this.calculateTextAttr(), {text: longestWord}))
+          .attr({...this.calculateTextAttr(), text: longestWord})
         this.cachedMinWidth = svgText.getBBox(true).width + this.padding.left + this.padding.right + 2 * this.getStroke()
         svgText.remove()
       }
@@ -129,7 +127,9 @@ draw2d.shape.basic.Text = draw2d.shape.basic.Label.extend(
 
       if (this.cachedWrappedAttr === null) {
         let abc = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-        let svgText = this.canvas.paper.text(0, 0, "").attr(extend({}, this.calculateTextAttr(), {text: abc}))
+        let svgText = this.canvas.paper.text(0, 0, "")
+        
+        svgText.attr({ ...this.calculateTextAttr(), text: abc})
 
         // get a good estimation of a letter width...not correct but this is working for the very first draft implementation
         let letterWidth = svgText.getBBox(true).width / abc.length
@@ -161,7 +161,6 @@ draw2d.shape.basic.Text = draw2d.shape.basic.Label.extend(
       }
       return this.cachedWrappedAttr
     }
-
   })
 
 
