@@ -262,7 +262,7 @@ draw2d.shape.pert.Activity = draw2d.shape.layout.VerticalLayout.extend(
     }
 
 
-    attributes = attributes || {}
+    attributes ??= {}
 
     if (this.getAlpha() < 0.9) {
       attributes.fill = this.bgColor.rgba()
