@@ -38,12 +38,18 @@ draw2d.util.spline.CubicSpline = draw2d.util.spline.Spline.extend(
     */
     generate: function(controlPoints, parts)
     {
-        // Endpoints are added twice to get them include in the
-        // generated array
-        var cp = new draw2d.util.ArrayList();
-        cp.add(controlPoints.get(0));
-        cp.addAll(controlPoints);
-        cp.add(controlPoints.get(controlPoints.getSize()-1));
+      // This implementation of an cubic spline needs at least 3 control points. Return the 
+      // original point if they do not meet them
+      if(n<3) {
+        return controlPoints.clone(true)
+      }
+
+      // Endpoints are added twice to get them include in the
+      // generated array
+      var cp = new draw2d.util.ArrayList();
+      cp.add(controlPoints.get(0));
+      cp.addAll(controlPoints);
+      cp.add(controlPoints.get(controlPoints.getSize()-1));
 
       var n = cp.getSize();
       var spline = new draw2d.util.ArrayList();
@@ -61,34 +67,35 @@ draw2d.util.spline.CubicSpline = draw2d.util.spline.Spline.extend(
     },
 
 
-      p: function( i,  t,  cp)
-      {
-        var x = 0.0;
-        var y = 0.0;
+    p: function( i,  t,  cp)
+    {
+      var x = 0.0;
+      var y = 0.0;
 
-        var k = i-1;
-        for (var j = -2; j <= 1; j++) {
-          var b = this.blend (j, t);
-          var p = cp.get(k++);
-          x += b * p.x;
-          y += b * p.y;
+      var k = i-1;
+      for (var j = -2; j <= 1; j++) {
+        var b = this.blend (j, t);
+        var p = cp.get(k++);
+        if(!p){
+          console.log("error")
         }
-
-       return new draw2d.geo.Point(x, y);
-      },
-
-
-
-      blend: function(i, t)
-      {
-        if (i === -2)
-            return (((-t + 3) * t - 3) * t + 1) / 6;
-        else if (i === -1)
-            return (((3 * t - 6) * t) * t + 4) / 6;
-        else if (i === 0)
-            return (((-3 * t + 3) * t + 3) * t + 1) / 6;
-
-        return (t * t * t) / 6;
+        x += b * p.x;
+        y += b * p.y;
       }
+
+      return new draw2d.geo.Point(x, y);
+    },
+
+    blend: function(i, t)
+    {
+      if (i === -2)
+          return (((-t + 3) * t - 3) * t + 1) / 6;
+      else if (i === -1)
+          return (((3 * t - 6) * t) * t + 4) / 6;
+      else if (i === 0)
+          return (((-3 * t + 3) * t + 3) * t + 1) / 6;
+
+      return (t * t * t) / 6;
+    }
 
 });
